feat: create a development environment with docker
Create a docker image with nodejs LTS and yarn and a tool to generate a license.
Create a docker-compose configuration to easily up a container and configure it.
Create a makefile based on the gazr method to manipulate the development
environment easily without knowledge about docker containers.
Generate a package.json to manage the project and create yarn commands,
these commands will be used by the makefile. All of these commands are not
implemented.
Complete the README to explain how to use this environment.