doc: add more info in README and complete the "how to start" instruction
complete the README to explain the server will run on localhost:9000 in
watch mode and the production bundle can be tested with an the run command
and will be available on localhost:3000.
add information about requirements and how to install them (it is
docker and docker-compose for now).

refactor the "how to start instruction" to be explicit on how to start
the development of the static website from scratch.
this little guide lists the files to update or delete to reinitialize the
website and start with a blank project.