feat: be able to run watch mode and run mode at the same time
the watch mode runs the webpack bundle with a live reload for development
on the port 9000.
the run mode runs the production bundle on port 3000.
due to the port reservation, the watch and the run command was not
able to be executed at the same time.
update the docker-compose so the watch mode container just
use the 9000 port.