feat: migrate to webpack 5
upgrade all packages and especially webpack and the associated packages
to the last major version.
migrate all configurations to be compatible with webpack 5.
Note: add the 'target: web' option in the dev config
to be able to have the hot reload.
it is due to a bug on webpack dev server and webpack 5.
upgrade the nodejs image with the new LTS version.