feat: deploy the bundle to GCP with terraform
add a Dockerfile with GCP CLI and terraform for the deployment.
create some tasks on the Makefile to easily deploy your production bundle
into a GCP bucket.
add terraform scripts to deploy an HTTPS static website to GCP.
add some documentation to describe the infrastructure.
add some documentation to guide the user to deploy his website.